Add AconfigPackageLocal
This change add AconfigPackageLocal. This class is intended only to be
used by generated code, and should be only used by aconfig team internal
to read the flag locally.

The class doesn't use the same implementation as AconfigPackage to
reduce the call stack, so that the execution is more efficient

Test: atest AconfigPackageTests
Bug: 367765164
Flag: NONE not api
Change-Id: I2876263630901dfe3113892467369e96a95692a6