Update allowlisted flags to include flags set during tests
Android 16 adds restrictions to the DeviceConfig flags that can
be set by the shell user, but a number of these flags are set by
adopting the shell permission identity during tests. This commit
updates the allowlist with all of the flags that were identified
as being set during tests.

Bug: 364083026
Flag: EXEMPT only adding allowlisted flags
Test: Manually ran a number of the tests for newly allowlisted flags
Change-Id: I76b33e0e74c1c326605f3470e47621c70887098a