Changed DeviceConfig.dump() signature
That method was taking a ParcelFileDescriptor and "requiring" DUMP
permission, but in reality it was just dumping internal state and not
using the file descriptor, so this change remove those.

In the long-term (or next CL :-), it might be even better to just
the listeners, and let the caller use them in a dump() method.

Test: m framework-configinfrastructure.stubs.source.module_lib-update-current-api
Test: atest ConfigInfrastructureServiceUnitTests:DeviceConfigTest
Fixes: 378760392
Bug: 364399200
Flag: android.provider.flags.dump_improvements

Change-Id: Ifab219bccededb07de558dced9d6203efe568666