Add static block in AconfigPackage
This change adds a static block in AconfigPackage. Thus when the class
is preloaded by zygote, it will load a packageName to PackageTable map,
and all processes can reuse this map to find the right PackageTable. It
saves the time for each process to iterate the folder to find the right
PackageTable.

Test: AconfigPackageTests
Bug: 349874828
Flag: NONE refactor
Change-Id: I45f06725538b57bd004143fb324cd6a4b3bd5d82