Remove device_config_overrides flags from allowlist
When a DeviceConfig flag is modified with the
DeviceConfig#setLocalOverride method, the original flag
is placed within the device_config_overrides namespace,
and this modified flag is used for an allowlist check. The
platform will filter out the device_config_overrides namespace
and use the original flag that was requested to be modified.
This commit removes any flags in this namespace and adds any
original flags from this namespace to the allowlist. Note,
some flags in this namespace are left to support the error
path scenarios in SettingsProviderTest.

Bug: 388901162
Flag: android.security.protect_device_config_flags
Test: atest DeviceConfigApiTests
Test: atest android.provider.SettingsProviderTest
Change-Id: Ie25a581d196d5f52f05d43ceb1c8a19ce62f2a56