Remove device_personalization_services flags from allowlist
Android 16 introduces a new restriction that only allowlisted
DeviceConfig flags can be modified by the shell user. The
device_personalization_services flags are not intended to be
modified by end users, so this commit removes all of the flags
in this namespace from the allowlist except for those that are
modified by tests adopting the shell permission identity.

Bug: 392611941
Flag: android.security.protect_device_config_flags
Test: atest PlatformScenarioTests
Change-Id: Ia27b5bf0aaca5efa7fa6f833687a79aab8d07c90